Calculate summary stats given a set of samples for instrumental forests.

Calculate summary stats given a set of samples for instrumental forests.

# S3 method for instrumental_forest
leaf_stats(forest, samples, ...)

Arguments

forest
The GRF forest

samples
The samples to include in the calculations.

...
Additional arguments (currently ignored).

Value

A named vector containing summary stats